working on issue #209
