+++ .github/workflows/pylint_check.yaml
@@ -81,5 +81,5 @@ jobs:
       - name: Running pylint
         run: cat action_pylint_files.txt | xargs python3 -m pylint --rcfile ./pylint.rc
       - name: Running pytest
-        run: pytest tests/test_rgb_indices.py
+        run: pytest test_rgb_indices.py
