+++ testing.py
@@ -7,7 +7,7 @@ import os
 import sys
 import numpy as np
 
-import gdal
+from osgeo import gdal
 
 import algorithm_rgb
